Allow groups to be placed as soon as they become useful
Previously, a group could only be placed if there were fewer than 17 accessible locations and if the group was useful. This was good for the Triforce Shards because it would mean that they would only get placed at the very end. However, since the number of accessible locations quickly goes above 16, it caused the pearls to not be placed until the end if they didn't become useful soon enough.