Item placement logic: Implement usefulness fractions
Instead of only distinguishing between items that would currently open
up a new location by itself (useful items) and items that wouldn't
(useless) we now calculate the fraction of how close an item is to being
useful. e.g. If you need 2 items to open a new location, both of those 2
items have 1/2 usefulness.

This helps a moderate amount in reducing the frequency of error messages
due to the randomizer logic dead-ending itself when only Dungeons (and
maybe a couple other locations) are selected for progress.