Dungeons only start mode: Prevent smalls keys in DRC big key chest
This is to prevent the item placement logic dead-ending itself.

Specifically, if the second small key in DRC got placed in the big key
chest, and no locations outside DRC are available, the randomizer would
be in a situation where it needs to place grappling hook and deku leaf
in the only two chests that are accessible. But if one of those two
chests got the DRC big key/dungeon map/compass placed in it beforehand,
it would be impossible to place both of those two items.