More aggressively move useless progress items to nonprogress
Previously an item would not be moved if it appeared anywhere in the
logic for any selected locations. Now items in the logic that are
obsoleted by a starting item are not included (e.g. Magic Meter Upgrade
in dungeons when you start with a sword).

This change also has the side effect of making usefulness fraction
calculation more accurate.