Allow user to specify a part of their starting inventory
This addresses LagoLunatic/wwrando#98 and LagoLunatic/wwrando#106

A new tab was introduced in the GUI to allow the user to select
their starting gear, mainly because it is too big to fit in the
current main tab.
Of course, if you don't use it, the randomizer behaves the same
as before.

A few implementation notes:
- This preallocates a 128 + 4 bytes array containing pointer to item_gets function, built from the item_get function array
- Progressive items are named as such in GUI for conveniance