Improve game loading speeds somewhat
Write file datas to the ISO in the same order they were written in vanilla, which is not the same order as the file entries. This improves performance of the game reading files for some reason.

Note that rando loading speeds are still slower than vanilla even with this change, just less so now.