Tweaks: Prevent removing entities from making entity indexes inconsistent
When various tweaks remove an entity, they previously actually deleted the entity from the list, but now they simply change its name to an empty string so that the original entity indexes are preserved. This allows enemy_locations.txt to always have accurate paths to enemies instead of the paths possibly being incorrect if a new tweak is added.