Fix bugs with the withered tree item
Fixed the code for changing how the withered tree item is spawned being very hacky and resulting in some random actor being treated like it was the item actor.
Also implemented giving the spawned item momentum so it flys up and over the player properly like it did in vanilla.