Fixed potential for inaccuracy when writing fixed-length strings
WW's fixed length strings (not counting magic identifiers) must still have a null byte at the end, so the actual max length of a fixed length string is one less than it appears. This fix prevents the possibility of writing strings that are so long they take up the last byte which should be reserved for the null terminator.