Simplify Hoho hint distribution code with modulo arithmetic
This fixes an error that would be thrown when the total number of hints available to all Hohos was less than 5.
It might also fix some other bugs, but I'm not sure as I can't understand the original code.
It also reduces the number of rng.shuffle calls from potentially as many as 100 to just 1.