Fix bad cave name keys
The cave keys in dungeon_and_cave_island_locations used when cave entrances aren't randomized were incorrect as they were using the entrance name, when they should be using the zone name.
This bug has existed for years but doesn't actually have any practical effect because it only happens when caves aren't randomized, and in that case they default to using the item location's zone name as the island name, which is always correct for caves.