Use deepcopy instead of copy
This fixes the potential for a major bug I introduced in 18fc8993515669e08a878b12c0acbb0c57a297e5 since item locations and macros are nested objects and are not properly copied by `copy`. If a macro was changed in one logic but not another, this could result in logic bugs.

I also use deepcopy for saving/loading the simulated playthrough state. This isn't currently necessary, but could theoretically be in the future.

deepcopy only adds about 0.1 seconds of hint generation times compared to copy.