Fix a bug in the randomizer's wordwrapping implementation
If a space occured exactly one character after the end of a line (e.g. "and that an item found at Tower of the Gods " is 44 characters including the space) then the function would add that space to the end of the line before creating a new line, forcing WW's text rendering engine to insert a blank line with nothing but a space on it.