Separate place/reward variables from display formatting
Some places in the code would try to check for the presence of location names in lists of Hint.place. This did not work when cryptic hints are enabled because in that case Hint.place contained the hint, not the location name.

So this separates the formatted hint text into separate variables, and leaves Hint.place and Hint.reward alone so those types of checks still work.