Implement proportional font wordwrapping
Previously, the randomizer would wordwrap dynamically-generated text as if it was using a monospace font, but Wind Waker actually uses a proportional font.
Now, the randomizer reads the font data and attempts to wordwrap text in exactly the same way as the game itself would, checking the width of every character in the font.