Make servant returned events end seamlessly
Previously, when the first servant got returned, its finish event would end, then switch_op would check the switch, then set its own switch and start an event. This gave the camera a frame or two to try to speed back to the player, which looked ugly.
That was neccessary due to a bug in the custom switch_op actor: it could only start events when no other event is running.

But now that switch_op has been refactored to work properly, it can be added to events, so that it can seamlessly start new events as soon as the previous one is finished.
This makes the camera instantly go to the tablet after the servant, with no attempt at zooming into the player.