Fix hints being inconsistent in dry runs
The hint randomizer was using WWRandomizer's RNG to generate hints instead of its own.
This caused the generated hints to be different depending on whether it was a dry run or a wet run, because the big pig color randomizer called reset_rng() an extra time in wet runs compared to dry runs.