Several improvements to entrance randomizer
The biggest change is that the nested entrance randomizer should no longer mix dungeons marked as required by race mode with dungeons marked as banned.

Dungeons-and-caves-only-start settings should now error out less frequently, if at all.

The code has been refactored to be a bit cleaner.

The race mode restriction of "multiple dungeons cannot appear on the same island" no longer applies to race-mode-banned dungeons. It should still prevent two required dungeons or a required dungeon and a banned dungeon from being on the same island, but that's it.