Fix errors when generating with dungeons-and-cave-only-start settings
Because of the way entrance rando works, a seed detected as "dungeons and caves only start" can effectively be a "dungeons only start" in reality because the caves are inaccessible. Thus all these hardcoded special exceptions for "dungeons only start" weren't triggering properly.

As a quick fix I just switched the special cases to check for "dungeons and caves only starts" too.

I'm pretty sure this is not a 100% perfect solution, but it does get rid of all of the errors and it's a lot easier than refactoring entrance randomizer again.