Fix yet another obscure racemode + entrance rando error case
In "dungeons only start" mode, combined with race mode, combined with "Nested Dungeons & Secret Caves (Together)" entrance randomizer, there was an edge case where an entrance chosen to lead into banned race mode dungeons was on the same island as the safety entrance chosen to prevent "dungeons only start" errors. Because of entrance-race-mode's special requirement of not allowing two dungeons on the same island unless both of them are banned due to overlapping quest markers, this scenario would result in a randomization error. This is now fixed.