Fix entrance rando error regarding FF
Forsaken Fortress is now no longer considered a special case for the entrance randomization logic, and is just a normal island entrance as far as it's concerned.
It should be able to handle an island entrance in the BOSS_ENTRANCES list now that the code has been generalized quite a bit.

This was necessary to fix errors when splitting progress and nonprogress entrances for unmixed caves and dungeons, as it couldn't find the last island entrance it needed.

Instead the hint randomizer considers FF a minor special case now.