Generalize dungeon progress subcategory handling
This generalizes the item rando's hardcoded exception for tingle chests so that it works with future non-progression locations being inside of progression dungeons.

Also explicitly commented race mode's behavior of forcing even the nonprogress locations within a dungeon to be accessible before you defeat that dungeon's boss, as this was not obvious with the argument to get_accessible_remaining_locations previously being implicit.