Started writing obscurity & precision level logic
Old existing comments on item locations about difficulty have been converted to macros, retested, and split into a gradation of 2*3 distinct levels.

Obscurity is for actions that are easy to do, but not obvious that they would be possible in the first place.
Precision is for actions that are hard to pull off even if you know what to do, things that may require good timing or aim.

Level 1 is just barely outside the bounds of what you'd expect from the vanilla game.
Level 2 is farther removed and may involve interactions the original developers didn't intend.
Level 3 is lunacy.