Improve custom actor adding patch
There was a bug causing the code to get into an infinite loop if you added more than one new custom actor.

Also, the number of new custom actors is now a constant in custom_data.asm that can be easily changed without having to modify a bunch of hardcoded +1 lines in misc_rando_features.asm.