Remove progress item groups
This feature was never properly coded, and it becomes more and more buggy with each new feature that gets added on top of it, causing errors, infinite loops, placing duplicate items, etc.

The original reason the feature was added was to allow the item randomizer's forward filler to be able to place progress items that aren't useful alone like triforce shards, but this was accomplished much more generally by the usefulness fractions feature that was added later.
The groups no longer seem to help the item randomizer generate seeds more consistently, and if anything might be causing the item randomizer to fail more frequently than with it removed.

Also, even when this feature worked correctly it resulted in randomized item placements that were noticeably biased to even casual players (e.g. getting all 8 shards as your first few items because the group was the first item placed).

This is a major change and might have unintended side effects somewhere.
The only one I've noticed is that the "Reward for All Tingle Statues" location is about 25% less likely to have progress items placed in it now (because the 5 items you need for it are no longer a group), but this doesn't seem like an issue.