Always use the Qt Fusion style for the GUI
This avoids an issue on Windows 11 where spinboxes wouldn't be wide enough to be readable when using the default style ("windows11").

Using several different styles on different operating systems and different color schemes makes it too hard to test and make sure the UI looks good and is usable on all platforms, so I'm just going to use this one cross-platform style for all of them instead.

Also fixed a few minor issues in the UI layout.