Prevent required and unrequired dungeons on the same island
In required bosses mode, when dungeon entrances are randomized,
but not nested boss entrances, we could end up placing a required boss
on the same island as a banned one, as we never randomize the boss
entrance itself (which is checked for co-location between banned and
required bosses), only the dungeon one.
Also ban required dungeons from appearing on islands with banned
dungeons. This additionally prevent confusion where a player might not
be able to tell which entrance can lead to progression if they don't
already know their bosses